chore: move codes from mod.rs to make more module
